Apply all the requested changes
