+++ app/src/main/java/com/epfl/beatlink/model/library/Playlist.kt
@@ -1,10 +1,8 @@
 package com.epfl.beatlink.model.library
 
-import com.epfl.beatlink.model.spotify.objects.SpotifyTrack
-
 data class Playlist(
     val playlistID: String,
-    val playlistCover: String, // TODO
+    val playlistCover: String,
     val playlistName: String, // mandatory
     val playlistDescription: String = "",
     val playlistPublic: Boolean = false,
@@ -19,9 +17,3 @@ data class Playlist(
     const val MAX_PLAYLIST_DESCRIPTION_LENGTH = 200
   }
 }
-
-data class PlaylistTrack(
-    val track: SpotifyTrack,
-    var likes: Int = 0,
-    val likedBy: MutableList<String> = mutableListOf()
-)
