+++ app/src/main/java/com/epfl/beatlink/model/library/PlaylistTrack.kt
@@ -1,0 +1,9 @@
+package com.epfl.beatlink.model.library
+
+import com.epfl.beatlink.model.spotify.objects.SpotifyTrack
+
+data class PlaylistTrack(
+    val track: SpotifyTrack,
+    var likes: Int = 0,
+    val likedBy: MutableList<String> = mutableListOf()
+)
