+++ app/src/main/java/com/epfl/beatlink/ui/components/search/DisplayResults.kt
@@ -1,0 +1,60 @@
+package com.epfl.beatlink.ui.components.search
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.unit.dp
+import com.epfl.beatlink.model.spotify.objects.SpotifyArtist
+import com.epfl.beatlink.model.spotify.objects.SpotifyTrack
+import com.epfl.beatlink.ui.components.library.TrackPlaylistItem
+import com.epfl.beatlink.viewmodel.library.PlaylistViewModel
+
+@Composable
+fun DisplayResults(
+    tracks: List<SpotifyTrack>? = null,
+    artists: List<SpotifyArtist>? = null,
+    playlistViewModel: PlaylistViewModel? = null,
+    onClearQuery: (() -> Unit)? = null
+) {
+  if (tracks.isNullOrEmpty() && artists.isNullOrEmpty()) {
+    // Empty state
+    Column(
+        modifier = Modifier.fillMaxSize().testTag("noResultsMessage"),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally) {
+          Text(
+              text = "No results found. Try searching for something.",
+              style = MaterialTheme.typography.bodyLarge)
+        }
+  } else {
+    // Display tracks or artists
+    LazyColumn(modifier = Modifier.testTag("searchResultsColumn")) {
+      tracks?.let {
+        items(it) { track ->
+          if (playlistViewModel != null && onClearQuery != null) {
+            TrackPlaylistItem(
+                track = track, playlistViewModel = playlistViewModel, onClearQuery = onClearQuery)
+          } else {
+            TrackItem(track = track)
+          }
+        }
+      }
+      artists?.let {
+        items(it) { artist ->
+          ArtistItem(artist = artist)
+          Spacer(modifier = Modifier.height(16.dp))
+        }
+      }
+    }
+  }
+}
